Rename and refactor music utility functions.
Renamed `utils.py` to `messages.py` to better reflect its role in handling user-facing messages. Updated message formatting with consistent emojis, colors, and added tips or links for improved user experience. Introduced an `age_restricted` embed and enhanced several existing embeds for clarity and usability.
